+++ lib/nacha/record/batch_header.rb
@@ -17,7 +17,7 @@ module Nacha
       nacha_field :company_entry_description, inclusion: 'M', contents: 'Alphameric', position: 54..63
       nacha_field :company_descriptive_date, inclusion: 'O', contents: 'Alphameric', position: 64..69
       nacha_field :effective_entry_date, inclusion: 'R', contents: 'YYMMDD', position: 70..75
-      nacha_field :settlement_date_julian, inclusion: 'M', contents: 'Numeric', position: 76..78
+      nacha_field :settlement_date_julian, inclusion: 'O', contents: 'Numeric', position: 76..78
       nacha_field :originator_status_code, inclusion: 'M', contents: 'Alphameric', position: 79..79
       nacha_field :originating_dfi_identification, inclusion: 'M', contents: 'TTTTAAAA', position: 80..87
       nacha_field :batch_number, inclusion: 'M', contents: 'Numeric', position: 88..94